API_Doc_QA_Portal/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pyyaml==6.0.2
pydantic==2.9.2
requests==2.32.3
pytest==8.3.3
openai==1.40.6
anthropic==0.34.2
sentence-transformers==3.0.1
chromadb==0.5.5
